chore(create-remix): move app creation to own module
